fix: follow-up bitcoin#15638 to move AnalyzePSBT related code
