+++ src/Makefile.am
@@ -259,6 +259,7 @@ BITCOIN_CORE_H = \
   node/coin.h \
   node/coinstats.h \
   node/context.h \
+  node/psbt.h \
   node/transaction.h \
   node/utxo_snapshot.h \
   noui.h \
@@ -468,6 +469,7 @@ libbitcoin_server_a_SOURCES = \
   node/coinstats.cpp \
   node/context.cpp \
   node/interfaces.cpp \
+  node/psbt.cpp \
   node/transaction.cpp \
   noui.cpp \
   policy/fees.cpp \
